+++ src/main/webapp/resources/js/components/AnalysesTable/AnalysisState.jsx
@@ -1,7 +1,8 @@
 import React from "react";
-import { Badge, Icon } from "antd";
+import { Badge } from "antd";
 import { SPACE_XS } from "../../styles/spacing";
 import { green6 } from "../../styles/colors";
+import { SyncOutlined } from "@ant-design/icons";
 
 /**
  * Display the state of an analysis
@@ -20,14 +21,10 @@ export function AnalysisState({ state }) {
       return <Badge status="success" text={state.text} />;
     default:
       return (
-        <span>
-          <Icon
-            type="sync"
-            style={{ marginRight: SPACE_XS, color: green6 }}
-            spin
-          />
+        <div>
+          <SyncOutlined style={{ marginRight: SPACE_XS, color: green6 }} spin />
           {state.text}
-        </span>
+        </div>
       );
   }
 }