+++ src/main/webapp/webpack.config.prod.js
@@ -36,7 +36,6 @@ exports.config = {
     ]
   },
   plugins: [
-    new webpack.ProgressPlugin(),
     new webpack.DefinePlugin({
       "process.env.NODE_ENV": JSON.stringify("production")
     }),