+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changes
 
 20.01 to 20.05
 --------------
+* [UI/Developer]: Removed webpack progress.
+* [UI]: Updated clients listing table to use `ant.design` tables.
+* [UI]: Sending email to sync user when a remote project's token expires.
 * [UI]: Fixed bug that prevented server side filtering for analysis name on analyses table. (20.01.01)
 
 19.09 to 20.01